[FLINK-36176] Remove support for Kafka 0.1
Warp up FLINK-19152 by also removing all related producer code.

Move FlinkKafkaProducerBase.getPropertiesFromBrokerList to a new test utility class as this method is used by some tests.
